Restructure graphql
