+++ src/resolvers/user.js
@@ -2,19 +2,6 @@ import { User as userCollection } from '../models/user'
 import mongoose from 'mongoose'
 import { prepare } from '../utils/index'
 
-export const typeDefs = `
-    extend type Query {
-        user(_id: String): User
-        users: [User]
-    }
-    type User {
-      _id: String,
-      email: String,
-      name: String,
-      username: String
-    }
-  `
-
 export const resolvers = {
   Query: {
     user: async (root, { _id }) => {
